More robust fallback to no apt pinning
Testing whether the explicitly specified packages are installable is not
sufficient, as the test bed might be in a state where installed versions are
already incompatible with the pinning. This happens when dist-upgrading the
testbed to the entirety of the specified pocket before setting up the pinning.

Now leave the apt pinning in place until we actually install test dependencies,
and retry install without pinning if that fails. This avoids the complicated
and insufficient installability heuristics and has the added benefit that this
grabs as many packages from the release pocket as possible. Thus adjust the
SchrootRunner.test_apt_pocket_pkg_with_proposed_dep test accordingly.

Document in the manpage that the pinning is not always successful and may fall
back to no pinning.