adt-virt-qemu: Implement support for nested base images
This adds initial support for nested base images to be passed into the
test environment, so that nested VMs may be used in tests.

A read-only copy of the first image without the overlay is passed to
the VM with a hardware serial number BASEIMAGE. adt-virt-qemu installs
udev rules that create a symbolic link /dev/baseimage for that drive
the first time the testbed is booted. Also, the symlink priority for
that drive is lowered, because the same file system UUIDs will be
present on both the first drive and the readonly baseimage drive.

The initramfs needs to be regenerated due to UUID collisions with the
baseimage drive, but we only need to do that if the testbed is rebooted.
So in turn just set a flag file and only regenerate the initramfs when a
reboot is requested.

Closes: #800845