Add simplified CLI parser
Create a new lib/autopkgtest_args.py which is a simplified version of the old
adt_run_args.py. It accepts only exactly one tested source package, but the
order of arguments is not relevant any more.

If no source package is given on the command line and the current working
directory is a Debian source package, test that.

This is preparation for LP: #1453509 and https://bugs.debian.org/795274 .