Add new "autopkgtest" CLI
Add runner/autopkgtest which provides a simplified CLI compared to the now
deprecated adt-run:

 - It accepts only exactly one tested source package
 - It defaults to the current directory if that is a Debian source package
 - The order of arguments is not relevant any more
 - Built vs. unbuilt tree is now automatically detected
 - The explicit --source, --click-source etc. options are gone, the type of
   tested source/binary packages is detected automatically

The old "adt-run" CLI will still be available for some time.

Create an adt-run → autopkgtest symlink for that which will import the old
adt_run_args.py module instead of the new autopkgtest_args.py (which provides
an identical API towards the caller). Once adt-run goes away,
autopkgtest_args.py and runner/autopkgtest can be simplified.

In runner/autopkgtest we now need to treat --testname and --override-control
differently: Take it directly from the options as "autopkgtest", but continue
to take it from the parsed actions as "adt-run".

Closes: #795274
LP: #1453509