Properly handle dist-upgrade failures in setup commands
If --setup-commands exits with code 100, consider this as a bad package
(autopkgtest exit code 12) instead of a transient testbed failure (exit code
16). This allows setup commands to detect errors which should be attributed to
the tested package itself (like dist-upgrading to the new version). Change
standard "apt-get update" setup commands to exit with code 1 to continue
treating that as a temporary failure.