Put virt servers back into $PATH
The virt servers were originally intended to be public API (even though not
documented as such) and other packages might provide virt servers. Put them
back into /usr/bin, with an "autopkgtest-virt-" prefix to mirror the original
"adt-virt-" naming.

Add <basedir>/virt/ to $PATH in autopkgtest so that running out of a git
checkout continues to work.

Rename the manpages to match the binaries.

Closes: #835204