ssh: Fix result of tests that break the testbed
If a test manages to break the testbed completly (kill sshd, networking, broken
kernel, etc.) and the test times out, we will also get ssh timeouts on trying
to fetch the logs and cleaning up.

In autopkgtest-virt-ssh's cleanup hook, ignore timeouts on cleaning actions in
the testbed. In Testbed.run_test(), ignore timeouts on copying up the logs and
skip the remaining post-processing if the test itself already timed out.

With that we get a proper "FAIL timed out" result with exit code 4 on these,
instead of a tmpfail/code 16 which would eternally loop on the same broken
test.

LP: #1630578