qemu: Hide detected udev file system properties
We don't want the OS to think that /dev/baseimage is a "real" file system, as
this thoroughly confuses e. g. btrfs when there are two block devices with the
same UUID. Extend our udev rule to null various ID_FS_* properties.

Also wait on udev reload to complete, to fix a race condition between creating
the rules and actually adding the base image. Thanks to Christian Seiler for
finding and testing this!

Closes: #842299