autopkgtest-virt-lxd: Check uptime for reboot waiting
I didn't quite get to the bottom of *why* this happens, but here's the
problem:

 - Some tests install a kernel
 - autopkgtest then reboots the testbed
 - After it comes back up, autopkgtest then carries on with the test

lxd's detection was using the fact that 'lxc exec ...' exits 0 if the
container is shut down to synchronise itself so that commands are
scheduled to the other side of the reboot. The problem is that this
wasn't working; the 'lxc exec' was exiting too early, autopkgtest was
carrying on and then the reboot actually happened when it was trying to
run tests.

As I say, I didn't manage to get to the bottom of the problem, but we
can fix the bug by switching to comparing the uptime directly. After the
reboot, the uptime should always be lower than it was before - watch for
this happening and then consider the container rebooted.

LP: #1654025