Add a debug-fail hook and implement it for autopkgtest-virt-ssh
At the minute, this is mainly so that the nova script can have its
failure information (`nova console-log') propagated up to the output, so
that in the case of kernel panics or other random failures we get useful
output that the driver of autopkgtest (e.g. autopkgtest-cloud) can look
at.

LP: #1630578