Use APT:Default-Release i.s.o. most pinning and add option --apt-default-release
To support the Debian use case of adding packages from sid/unstable to testing,
we want to be independent of the fact that unstable can be called by its code
name or by the suite (as adding code that checks for the delta feels clumsy and
looks like a hack). To make sure that one can fix malfunctioning autodetection,
make it an option to the outside.

A bonus of using APT:Default-Release is the reduced amount of pinning
needed. For this to be equivalent, the pinning values that remain are changed
to APT defaults.

Because the default release is needed in multiple places, it is now a separate
function, that uses caching.