Fix release name parsing with arch-qualified sources.list
On my system I have a mixture of arch-qualified (deb) and not-arch-qualified
(deb-src) lines in sources.list.  This results in 'autopkgtest
--apt-pocket=proposed=src:perl' trying to create files with names like
'/etc/apt/preferences.d/autopkgtest-http://archive.ubuntu.com/ubuntu-proposed'.

It's not obvious to me why the sed script in question behaves this way, but
it appears that e.g. this line is matching when it shouldn't and resulting
in the broken output:

  deb [arch=amd64,i386] http://archive.ubuntu.com/ubuntu artful-updates main

Fix this a two-pass sed script, first ignoring the [] field and then applying
the rest of the rules to extract the series name.

Closes: #870393

Gbp-Dch: Short